style: apply safe readability cleanups

+++ dialogs/MRAboutDialog.cpp
@@ -37,7 +37,7 @@ TRect centeredRect(int width, int height) {
 }
 
 void insertCenteredStaticLine(TDialog *dialog, int width, int y, const std::string &text) {
-	int x = std::max(2, (width - static_cast<int>(strwidth(text.c_str()))) / 2);
+	int x = std::max(2, (width - strwidth(text.c_str())) / 2);
 	dialog->insert(new TStaticText(TRect(x, y, width - 2, y + 1), text.c_str()));
 }
